Working bad comented libraries
Implemented new library utils where low lever function I think should be
implemented. For example to working with bytes of a packet and packet
Manage library to use the loop.

The tcp constructor is working and now can extract the parts of a tcp
and allocate them in a TCP_packet struct.

This is useful because FTP only uses port 21.

Also another function is implemented that prints the payload of a TCP
packet. Must test in a FTP packet(not tested yet).

TODO: work with ARP packets and see how to send them to poison.